// src/client/types.gen.ts
export type ApiResponseCharacterDetailOut = {
    /**
     * Status
     *
     * 响应状态
     */
    status?: string;
    /**
     * 响应数据
     */
    data?: CharacterDetailOut | null;
    /**
     * Message
     *
     * 响应信息
     */
    message?: string;
    /**
     * Error Code
     *
     * 内部错误码
     */
    error_code?: string | null;
    /**
     * Timestamp
     *
     * 响应时间戳
     */
    timestamp?: number;
};
// ...
export type CharacterDetailOut = {
    /**
     * Name
     *
     * 角色名称
     */
    name: string;
    /**
     * Avatar
     *
     * 角色头像
     */
    avatar?: string | null;
    /**
     * Bio
     *
     * 角色个性签名
     */
    bio?: string | null;
    /**
     * Tags
     *
     * 角色标签
     */
    tags?: Array<string>;
    /**
     * Ispublic
     *
     * 是否公开
     */
    isPublic?: boolean;
    /**
     * Status
     *
     * 角色状态
     */
    status?: 'draft' | 'active' | 'pending' | 'deleted' | 'banned';
    /**
     * Id
     *
     * 角色ID
     */
    id: string;
    /**
     * Authorid
     *
     * 作者ID
     */
    authorId: string;
    /**
     * Systemprompt
     *
     * 系统提示词
     */
    systemPrompt?: string;
    /**
     * Baseconfig
     *
     * LLM生成参数
     */
    baseConfig?: {
        [key: string]: unknown;
    };
    /**
     * Favoritecount
     *
     * 总收藏数
     */
    favoriteCount: number;
    /**
     * Chatcount
     *
     * 聊天数
     */
    chatCount: number;
    /**
     * Isfavorited
     *
     * 当前用户是否已收藏
     */
    isFavorited?: boolean;
    /**
     * Createdat
     *
     * 创建时间
     */
    createdAt: string;
    /**
     * Updatedat
     *
     * 更新时间
     */
    updatedAt: string;
};